+++ Services/Ecs/V2/Model/ListServersDetailsRequest.cs
@@ -37,7 +37,7 @@ namespace HuaweiCloud.SDK.Ecs.V2.Model
         public string Ip { get; set; }
 
         /// <summary>
-        /// 查询返回云服务器当前页面的大小。每页最多返回1000台云服务器的信息。
+        /// 查询返回云服务器当前页面的大小。每页默认值是25，最多返回1000台云服务器的信息。
         /// </summary>
         [SDKProperty("limit", IsQuery = true)]
         [JsonProperty("limit", NullValueHandling = NullValueHandling.Ignore)]