+++ Services/GaussDBforNoSQL/V3/Model/UpdateSecurityGroupResponse.cs
@@ -16,7 +16,7 @@ namespace HuaweiCloud.SDK.GaussDBforNoSQL.V3.Model
     {
 
         /// <summary>
-        /// 工作流ID。
+        /// 任务ID。
         /// </summary>
         [JsonProperty("job_id", NullValueHandling = NullValueHandling.Ignore)]
         public string JobId { get; set; }