+++ Services/Image/V2/Model/ImageDescriptionReq.cs
@@ -22,16 +22,10 @@ namespace HuaweiCloud.SDK.Image.V2.Model
         public string Image { get; set; }
 
         /// <summary>
-        /// 与image二选一  图片的URL路径，目前支持：  - 公网HTTP/HTTPS URL  - 华为云OBS提供的URL，使用OBS数据需要进行授权。包括对服务授权、临时授权、匿名公开授权。详请参见[配置OBS服务的访问权限](https://support.huaweicloud.com/api-moderation/moderation_03_0020.html)。  &gt; - 接口响应时间依赖于图片的下载时间，如果图片下载时间过长，会返回接口调用失败。 &gt; - 请保证被检测图片所在的存储服务稳定可靠，建议您使用华为云OBS存储。 
+        /// 与image二选一  图片的URL路径，目前支持：  - 公网HTTP/HTTPS URL  - 华为云OBS提供的URL，使用OBS数据需要进行授权。包括对服务授权、临时授权、匿名公开授权。详请参见[配置OBS服务的访问权限](https://support.huaweicloud.com/api-image/image_03_0037.html)。  &gt; - 接口响应时间依赖于图片的下载时间，如果图片下载时间过长，会返回接口调用失败。 &gt; - 请保证被检测图片所在的存储服务稳定可靠，建议您使用华为云OBS存储。 
         /// </summary>
         [JsonProperty("url", NullValueHandling = NullValueHandling.Ignore)]
         public string Url { get; set; }
-
-        /// <summary>
-        /// zh：返回标签的语言类型为中文。 en：返回标签的语言类型为英文。 默认值为en。 
-        /// </summary>
-        [JsonProperty("language", NullValueHandling = NullValueHandling.Ignore)]
-        public string Language { get; set; }
 
 
         /// <summary>
@@ -43,7 +37,6 @@ namespace HuaweiCloud.SDK.Image.V2.Model
             sb.Append("class ImageDescriptionReq {\n");
             sb.Append("  image: ").Append(Image).Append("\n");
             sb.Append("  url: ").Append(Url).Append("\n");
-            sb.Append("  language: ").Append(Language).Append("\n");
             sb.Append("}\n");
             return sb.ToString();
         }
@@ -74,11 +67,6 @@ namespace HuaweiCloud.SDK.Image.V2.Model
                     this.Url == input.Url ||
                     (this.Url != null &&
                     this.Url.Equals(input.Url))
-                ) && 
-                (
-                    this.Language == input.Language ||
-                    (this.Language != null &&
-                    this.Language.Equals(input.Language))
                 );
         }
 
@@ -94,8 +82,6 @@ namespace HuaweiCloud.SDK.Image.V2.Model
                     hashCode = hashCode * 59 + this.Image.GetHashCode();
                 if (this.Url != null)
                     hashCode = hashCode * 59 + this.Url.GetHashCode();
-                if (this.Language != null)
-                    hashCode = hashCode * 59 + this.Language.GetHashCode();
                 return hashCode;
             }
         }