+++ Services/ProjectMan/V4/Model/ListProjectDomainsResponse.cs
@@ -1,0 +1,90 @@
+using System;
+using System.Collections.Generic;
+using System.Text;
+using System.Linq;
+using System.Runtime.Serialization;
+using Newtonsoft.Json;
+using Newtonsoft.Json.Converters;
+using HuaweiCloud.SDK.Core;
+
+namespace HuaweiCloud.SDK.ProjectMan.V4.Model
+{
+    /// <summary>
+    /// Response Object
+    /// </summary>
+    public class ListProjectDomainsResponse : SdkResponse
+    {
+
+        /// <summary>
+        /// 领域总数
+        /// </summary>
+        [JsonProperty("total", NullValueHandling = NullValueHandling.Ignore)]
+        public int? Total { get; set; }
+
+        /// <summary>
+        /// 领域列表
+        /// </summary>
+        [JsonProperty("domains", NullValueHandling = NullValueHandling.Ignore)]
+        public List<CreateProjectDomainResponseBody> Domains { get; set; }
+
+
+        /// <summary>
+        /// Get the string
+        /// </summary>
+        public override string ToString()
+        {
+            var sb = new StringBuilder();
+            sb.Append("class ListProjectDomainsResponse {\n");
+            sb.Append("  total: ").Append(Total).Append("\n");
+            sb.Append("  domains: ").Append(Domains).Append("\n");
+            sb.Append("}\n");
+            return sb.ToString();
+        }
+
+        /// <summary>
+        /// Returns true if objects are equal
+        /// </summary>
+        public override bool Equals(object input)
+        {
+            return this.Equals(input as ListProjectDomainsResponse);
+        }
+
+        /// <summary>
+        /// Returns true if objects are equal
+        /// </summary>
+        public bool Equals(ListProjectDomainsResponse input)
+        {
+            if (input == null)
+                return false;
+
+            return 
+                (
+                    this.Total == input.Total ||
+                    (this.Total != null &&
+                    this.Total.Equals(input.Total))
+                ) && 
+                (
+                    this.Domains == input.Domains ||
+                    this.Domains != null &&
+                    input.Domains != null &&
+                    this.Domains.SequenceEqual(input.Domains)
+                );
+        }
+
+        /// <summary>
+        /// Get hash code
+        /// </summary>
+        public override int GetHashCode()
+        {
+            unchecked // Overflow is fine, just wrap
+            {
+                int hashCode = 41;
+                if (this.Total != null)
+                    hashCode = hashCode * 59 + this.Total.GetHashCode();
+                if (this.Domains != null)
+                    hashCode = hashCode * 59 + this.Domains.GetHashCode();
+                return hashCode;
+            }
+        }
+    }
+}