+++ Services/ProjectMan/V4/ProjectManAsyncClient.cs
@@ -530,6 +530,25 @@ namespace HuaweiCloud.SDK.ProjectMan.V4
         }
         
         /// <summary>
+        /// 取消领域与项目的关联关系
+        ///
+        /// 取消领域与项目的关联关系
+        /// 
+        /// 详细说明请参考华为云API Explorer。
+        /// Please refer to Huawei cloud API Explorer for details.
+        /// </summary>
+        public async Task<CancelProjectDomainResponse> CancelProjectDomainAsync(CancelProjectDomainRequest cancelProjectDomainRequest)
+        {
+            Dictionary<string, string> urlParam = new Dictionary<string, string>();
+            urlParam.Add("project_id" , cancelProjectDomainRequest.ProjectId.ToString());
+            urlParam.Add("domain_id" , cancelProjectDomainRequest.DomainId.ToString());
+            string urlPath = HttpUtils.AddUrlPath("/v4/projects/{project_id}/domains/{domain_id}",urlParam);
+            SdkRequest request = HttpUtils.InitSdkRequest(urlPath, "application/json", cancelProjectDomainRequest);
+            HttpResponseMessage response = await DoHttpRequestAsync("DELETE",request);
+            return JsonUtils.DeSerializeNull<CancelProjectDomainResponse>(response);
+        }
+        
+        /// <summary>
         /// 创建工作项类型自定义字段
         ///
         /// 创建工作项类型自定义字段
@@ -581,6 +600,24 @@ namespace HuaweiCloud.SDK.ProjectMan.V4
             SdkRequest request = HttpUtils.InitSdkRequest(urlPath, "application/json", createIterationV4Request);
             HttpResponseMessage response = await DoHttpRequestAsync("POST",request);
             return JsonUtils.DeSerialize<CreateIterationV4Response>(response);
+        }
+        
+        /// <summary>
+        /// 创建项目的领域
+        ///
+        /// 查询项目的领域列表
+        /// 
+        /// 详细说明请参考华为云API Explorer。
+        /// Please refer to Huawei cloud API Explorer for details.
+        /// </summary>
+        public async Task<CreateProjectDomainResponse> CreateProjectDomainAsync(CreateProjectDomainRequest createProjectDomainRequest)
+        {
+            Dictionary<string, string> urlParam = new Dictionary<string, string>();
+            urlParam.Add("project_id" , createProjectDomainRequest.ProjectId.ToString());
+            string urlPath = HttpUtils.AddUrlPath("/v4/projects/{project_id}/domain",urlParam);
+            SdkRequest request = HttpUtils.InitSdkRequest(urlPath, "application/json", createProjectDomainRequest);
+            HttpResponseMessage response = await DoHttpRequestAsync("POST",request);
+            return JsonUtils.DeSerialize<CreateProjectDomainResponse>(response);
         }
         
         /// <summary>
@@ -882,6 +919,24 @@ namespace HuaweiCloud.SDK.ProjectMan.V4
         }
         
         /// <summary>
+        /// 查询项目的领域列表
+        ///
+        /// 查询项目的领域列表
+        /// 
+        /// 详细说明请参考华为云API Explorer。
+        /// Please refer to Huawei cloud API Explorer for details.
+        /// </summary>
+        public async Task<ListProjectDomainsResponse> ListProjectDomainsAsync(ListProjectDomainsRequest listProjectDomainsRequest)
+        {
+            Dictionary<string, string> urlParam = new Dictionary<string, string>();
+            urlParam.Add("project_id" , listProjectDomainsRequest.ProjectId.ToString());
+            string urlPath = HttpUtils.AddUrlPath("/v4/projects/{project_id}/domains",urlParam);
+            SdkRequest request = HttpUtils.InitSdkRequest(urlPath, "application/json", listProjectDomainsRequest);
+            HttpResponseMessage response = await DoHttpRequestAsync("GET",request);
+            return JsonUtils.DeSerialize<ListProjectDomainsResponse>(response);
+        }
+        
+        /// <summary>
         /// 查询项目下所有工作项的历史记录
         ///
         /// 查询项目下所有工作项的历史记录
@@ -1079,6 +1134,25 @@ namespace HuaweiCloud.SDK.ProjectMan.V4
             SdkRequest request = HttpUtils.InitSdkRequest(urlPath, "application/json", updateIterationV4Request);
             HttpResponseMessage response = await DoHttpRequestAsync("PUT",request);
             return JsonUtils.DeSerializeNull<UpdateIterationV4Response>(response);
+        }
+        
+        /// <summary>
+        /// 更新项目的领域
+        ///
+        /// 更新项目的领域
+        /// 
+        /// 详细说明请参考华为云API Explorer。
+        /// Please refer to Huawei cloud API Explorer for details.
+        /// </summary>
+        public async Task<UpdateProjectDomainResponse> UpdateProjectDomainAsync(UpdateProjectDomainRequest updateProjectDomainRequest)
+        {
+            Dictionary<string, string> urlParam = new Dictionary<string, string>();
+            urlParam.Add("project_id" , updateProjectDomainRequest.ProjectId.ToString());
+            urlParam.Add("domain_id" , updateProjectDomainRequest.DomainId.ToString());
+            string urlPath = HttpUtils.AddUrlPath("/v4/projects/{project_id}/domains/{domain_id}",urlParam);
+            SdkRequest request = HttpUtils.InitSdkRequest(urlPath, "application/json", updateProjectDomainRequest);
+            HttpResponseMessage response = await DoHttpRequestAsync("PUT",request);
+            return JsonUtils.DeSerialize<UpdateProjectDomainResponse>(response);
         }
         
         /// <summary>